requirements.txt
absl-py==0.9.0
adal==1.2.2
alabaster==0.7.12
albumentations==0.4.5
anaconda-client==1.7.2
anaconda-project==0.8.3
ansiwrap==0.8.4
appdirs==1.4.3
archspec==0.1.1
argh==0.26.2
arrow==0.15.5
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0.1.post1
async-generator==1.10
atari-py==0.2.6
atomicwrites==1.3.0
attrs==19.3.0
audioread==2.1.8
autopep8==1.5.2
Babel==2.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcolz==1.2.1
beautifulsoup4==4.9.0
binaryornot==0.4.4
bitarray==1.2.1
bkcharts==0.2
black==19.10b0
bleach==3.1.4
blinker==1.4
blis==0.2.4
bokeh==2.0.1
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==3.1.1
catboost==0.23.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
conda==4.8.4
conda-package-handling==1.6.0
confuse==1.1.0
contextlib2==0.6.0.post1
cookiecutter==1.7.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
Cython==0.29.16
cytoolz==0.10.1
dask==2.15.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.15.0
docker==4.2.0
docker-pycreds==0.4.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastai==1.0.60
fastcache==1.1.0
fastprogress==0.2.2
ffmpeg-python==0.2.0
filelock==3.0.10
fire==0.1.3
flake8==3.7.9
Flask==1.1.2
fsspec==0.7.3
future==0.18.2
gast==0.3.3
gcsfs==0.6.1
gevent==20.4.0
gitdb==4.0.4
GitPython==3.1.1
glob2==0.7
gmpy2==2.1.0b1
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigtable==1.2.1
google-cloud-core==1.3.0
google-cloud-dataproc==0.7.0
google-cloud-datastore==1.12.0
google-cloud-firestore==1.6.2
google-cloud-kms==1.4.0
google-cloud-language==1.3.0
google-cloud-logging==1.15.0
google-cloud-pubsub==1.4.3
google-cloud-scheduler==1.2.1
google-cloud-spanner==1.15.1
google-cloud-speech==1.3.2
google-cloud-storage==1.28.0
google-cloud-tasks==1.5.0
google-cloud-translate==2.0.1
google-cloud-videointelligence==1.14.0
google-cloud-vision==1.0.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.14
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.23.0
grpcio-gcp==0.2.2
gym==0.17.2
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
htmlmin==0.1.12
httplib2==0.17.3
hypothesis==5.10.4
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imgaug==0.2.6
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.2.1
ipynb-path==0.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.15.1
json5==0.9.0
jsonschema==3.2.0
-e git+https://github.com/openai/jukebox.git@830f9b65cff9c346e7bee4c97ec049fe139447f2#egg=jukebox
jupyter-aihub-deploy-extension==0.1
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-http-over-ws==0.0.8
jupyterlab==2.1.4
jupyterlab-git==0.20.0
jupyterlab-server==1.1.1
kaggle==1.5.6
kaggle-environments==0.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.1.1
kiwisolver==1.2.0
kubernetes==10.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
librosa==0.7.2
lief==0.9.0
lightgbm==2.3.1
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mcs-kfold==0.1.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
monai==0.3.0
more-itertools==8.2.0
mpi4py==3.0.3
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
munch==2.5.0
murmurhash==1.0.0
mypy-extensions==0.4.3
nb-conda==2.2.1
nb-conda-kernels==2.2.3
nbclient==0.2.0
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.6
nbpresent==3.0.2
nest-asyncio==1.3.2
networkx==2.4
nibabel==3.1.0
nilearn==0.6.2
nltk==3.4.4
norbert==0.2.1
nose==1.3.7
notebook==6.0.3
notebook-executor==0.2
numba==0.48.0
numexpr==2.7.1
numpy==1.19.1
numpydoc==0.9.2
oauthlib==3.0.1
olefile==0.46
opencv-python==4.3.0.36
opencv-python-headless==4.2.0.34
openpyxl==3.0.3
opt-einsum==0+untagged.61.gd905544.dirty
packaging==20.1
pandas==1.1.1
pandas-profiling==2.6.0
pandocfilters==1.4.2
papermill==2.1.0
parso==0.5.2
partd==1.1.0
path==13.2.0
path.py==12.0.2
pathlib2==2.3.5
pathspec==0.8.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phik==0.9.11
pickleshare==0.7.5
Pillow==7.1.1
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.8.2
pluggy==0.13.0
ply==3.11
pooch==1.1.1
poyo==0.5.0
preshed==2.0.1
pretrainedmodels==0.7.4
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
pydicom==2.0.0
pydocstyle==5.0.2
pydot==1.4.1
pyflakes==2.1.1
pyglet==1.5.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-slugify==4.0.0
pytorch-tabnet==1.1.0
pytorchcv==0.0.58
pytz==2019.3
PyWavelets==1.1.1
pyworld==0.2.10
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.0
QDarkStyle==2.8.1
QtAwesome==0.7.1
qtconsole==4.7.3
QtPy==1.9.0
RandAugment==0.1
regex==2019.8.19
requests==2.23.0
requests-oauthlib==1.2.0
resampy==0.2.2
retrying==1.3.3
rope==0.16.0
rsa==4.0
Rtree==0.9.4
ruamel-yaml==0.15.80
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
sip==4.19.20
six==1.14.0
sklearn==0.0
smmap==3.0.2
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
SoundFile==0.10.3.post1
soupsieve==1.9.4
spacy==2.1.8
Sphinx==3.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.1
spleeter==1.5.3
spyder==4.1.2
spyder-kernels==1.9.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
srsly==0.1.0
stable-baselines==2.9.0
statsmodels==0.11.1
sympy==1.5.1
tables==3.6.1
tangled-up-in-unicode==0.0.4
tblib==1.6.0
tenacity==6.1.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
textwrap3==0.9.2
thinc==7.0.8
threadpoolctl==2.1.0
timm==0.1.26
toml==0.10.0
toolz==0.10.0
torc==0.1.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.24.3
visions==0.4.1
warmup-scheduler==0.3.2
wasabi==0.2.2
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xgboost==1.1.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
yapf==0.29.0
zict==2.0.0
zipp==3.1.0